Moved Game Role classes to gamerole package
